Sonar whenn

+++ src/creatures/players/player.cpp
@@ -3516,7 +3516,7 @@ void Player::removeThing(Thing* thing, uint32_t count)
 
 uint8_t Player::getThingIndex(const Thing* thing) const
 {
-	for (int i = CONST_SLOT_FIRST; i <= CONST_SLOT_LAST; ++i) {
+	for (uint8_t i = CONST_SLOT_FIRST; i <= CONST_SLOT_LAST; ++i) {
 		if (inventory[i] == thing) {
 			return i;
 		}
