+++ src/items/containers/container.cpp
@@ -684,7 +684,7 @@ void Container::removeThing(Thing* thing, uint32_t count)
 
 uint8_t Container::getThingIndex(const Thing* thing) const
 {
-	int32_t index = 0;
+	uint8_t index = 0;
 	for (Item* item : itemlist) {
 		if (item == thing) {
 			return index;
